+++ app/src/main/java/com/github/mobile/ui/user/UserViewActivity.java
@@ -18,6 +18,8 @@ package com.github.mobile.ui.user;
 import static android.content.Intent.FLAG_ACTIVITY_CLEAR_TOP;
 import static android.content.Intent.FLAG_ACTIVITY_SINGLE_TOP;
 import static com.github.mobile.Intents.EXTRA_USER;
+
+import android.accounts.Account;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
@@ -33,11 +35,14 @@ import com.github.mobile.core.user.FollowUserTask;
 import com.github.mobile.core.user.FollowingUserTask;
 import com.github.mobile.core.user.RefreshUserTask;
 import com.github.mobile.core.user.UnfollowUserTask;
+import com.github.mobile.core.user.UserComparator;
+import com.github.mobile.persistence.AccountDataManager;
 import com.github.mobile.ui.TabPagerActivity;
 import com.github.mobile.util.AvatarLoader;
 import com.github.mobile.util.ToastUtils;
 import com.github.mobile.util.TypefaceUtils;
 import com.google.inject.Inject;
+import com.google.inject.Provider;
 
 import org.eclipse.egit.github.core.User;
 
@@ -58,11 +63,21 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
     }
 
     @Inject
+    private AccountDataManager accountDataManager;
+
+    @Inject
+    private Provider<UserComparator> userComparatorProvider;
+
+    @Inject
     private AvatarLoader avatars;
 
     private User user;
 
     private ProgressBar loadingBar;
+
+    private boolean isOrganization;
+
+    private boolean isMember;
 
     private boolean isFollowing;
 
@@ -79,12 +94,22 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
         actionBar.setDisplayHomeAsUpEnabled(true);
         actionBar.setTitle(user.getLogin());
 
-        if (!TextUtils.isEmpty(user.getAvatarUrl()) && !TextUtils.isEmpty(user.getType()))
+        if (!TextUtils.isEmpty(user.getAvatarUrl()) && User.TYPE_USER.equals(user.getType())) {
             configurePager();
-        else {
+        } else {
             ViewUtils.setGone(loadingBar, false);
             setGone(true);
             new RefreshUserTask(this, user.getLogin()) {
+                @Override
+                protected User run(Account account) throws Exception {
+                    isMember = false;
+                    for (User org : accountDataManager.getOrgs(false)) {
+                        if (user.getLogin().equals(org.getLogin())) {
+                            isMember = true;
+                        }
+                    }
+                    return super.run(account);
+                }
 
                 @Override
                 protected void onSuccess(User fullUser) throws Exception {
@@ -109,7 +134,7 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
 
     @Override
     public boolean onCreateOptionsMenu(Menu optionsMenu) {
-        if (!isOrganization()) {
+        if (!isOrganization) {
             getMenuInflater().inflate(R.menu.user_follow, optionsMenu);
         }
 
@@ -118,7 +143,7 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
 
     @Override
     public boolean onPrepareOptionsMenu(Menu menu) {
-        if (!isOrganization()) {
+        if (!isOrganization) {
             MenuItem followItem = menu.findItem(R.id.m_follow);
 
             followItem.setVisible(followingStatusChecked);
@@ -147,6 +172,7 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
 
     private void configurePager() {
         avatars.bind(getSupportActionBar(), user);
+        isOrganization = User.TYPE_ORG.equals(user.getType());
         configureTabPager();
         ViewUtils.setGone(loadingBar, true);
         setGone(false);
@@ -166,7 +192,7 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
 
     @Override
     protected UserPagerAdapter createAdapter() {
-        return new UserPagerAdapter(this, isOrganization());
+        return new UserPagerAdapter(this, isOrganization, isMember);
     }
 
     @Override
@@ -242,9 +268,5 @@ public class UserViewActivity extends TabPagerActivity<UserPagerAdapter>
                 invalidateOptionsMenu();
             }
         }.execute();
-    }
-
-    private boolean isOrganization() {
-        return User.TYPE_ORG.equals(user.getType());
     }
 }